Add a monotonically incrementing persistent id counter
When a persistent ID is not provided, we were generating a "unique" one
by constructing a string with the current seconds and microseconds.

If two connections are createed in rapid succession, they can both
generate the same `persistent_id` and therefore both use the same
underlying `php_stream`.

```php
$r1 = new Redis;
$r2 = new Redis;

// If the two of these execute on the same second and usec, they will both use
// the same persistent_id and therefore the same underlying socket.
$r1->pconnect('localhost', 6379);
$r2->pconnect('localhost', 6379);

// Close and free the first connection
$r1->close();

// Segfault. Closing an already closed stream
$r2->close();
```

This commit simply adds a function scoped static monotonically
incremented counter which is used to make the persistent id's truly unique.

Fixes #2762