docs: Fix `@see \Redis` annotations so they hyperlink correctly.
It turns out `doctum` will only hyperlink to the Redis methods when the
method casing is identical. For that reason, just match the casing in
`redis_cluster.stub.php` to how it's cased in `redis.stub.php`.

We aren't changing any of the non-commented stub text, just the casing of the
`@see` annotations.

WIP: More

.